MetricDao.selectNullableById should really take a long, not a int

+++ server/sonar-server/src/main/java/org/sonar/server/metric/persistence/MetricDao.java
@@ -114,7 +114,7 @@ public class MetricDao implements DaoComponent {
     mapper(session).update(metric);
   }
 
-  public MetricDto selectNullableById(DbSession session, int id) {
+  public MetricDto selectNullableById(DbSession session, long id) {
     return mapper(session).selectById(id);
   }
 }
